+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## 0.3.8 - 2017-08-20
+
+### Fixed
+- many codacy errors
+
+## 0.3.7 - 2017-08-20
+
 ### Added
 - add login menu option in blogs view
 - more SpecTests